+++ sdk/resourcemanager/extendedlocation/armextendedlocation/models_serde.go
@@ -16,9 +16,40 @@ import (
 	"reflect"
 )
 
+// MarshalJSON implements the json.Marshaller interface for type AwsCloudProfile.
+func (a AwsCloudProfile) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "accountId", a.AccountID)
+	populate(objectMap, "excludedAccounts", a.ExcludedAccounts)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AwsCloudProfile.
+func (a *AwsCloudProfile) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "accountId":
+			err = unpopulate(val, "AccountID", &a.AccountID)
+			delete(rawMsg, key)
+		case "excludedAccounts":
+			err = unpopulate(val, "ExcludedAccounts", &a.ExcludedAccounts)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type CustomLocation.
 func (c CustomLocation) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", c.ID)
 	populate(objectMap, "identity", c.Identity)
 	populate(objectMap, "location", c.Location)
@@ -73,7 +104,7 @@ func (c *CustomLocation) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationFindTargetResourceGroupProperties.
 func (c CustomLocationFindTargetResourceGroupProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "labels", c.Labels)
 	return json.Marshal(objectMap)
 }
@@ -100,7 +131,7 @@ func (c *CustomLocationFindTargetResourceGroupProperties) UnmarshalJSON(data []b
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationFindTargetResourceGroupResult.
 func (c CustomLocationFindTargetResourceGroupResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "matchedResourceSyncRule", c.MatchedResourceSyncRule)
 	populate(objectMap, "targetResourceGroup", c.TargetResourceGroup)
 	return json.Marshal(objectMap)
@@ -131,7 +162,7 @@ func (c *CustomLocationFindTargetResourceGroupResult) UnmarshalJSON(data []byte)
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationListResult.
 func (c CustomLocationListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", c.NextLink)
 	populate(objectMap, "value", c.Value)
 	return json.Marshal(objectMap)
@@ -162,7 +193,7 @@ func (c *CustomLocationListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationOperation.
 func (c CustomLocationOperation) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "display", c.Display)
 	populate(objectMap, "isDataAction", c.IsDataAction)
 	populate(objectMap, "name", c.Name)
@@ -201,7 +232,7 @@ func (c *CustomLocationOperation) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationOperationValueDisplay.
 func (c CustomLocationOperationValueDisplay) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "description", c.Description)
 	populate(objectMap, "operation", c.Operation)
 	populate(objectMap, "provider", c.Provider)
@@ -240,7 +271,7 @@ func (c *CustomLocationOperationValueDisplay) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationOperationsList.
 func (c CustomLocationOperationsList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", c.NextLink)
 	populate(objectMap, "value", c.Value)
 	return json.Marshal(objectMap)
@@ -271,10 +302,12 @@ func (c *CustomLocationOperationsList) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationProperties.
 func (c CustomLocationProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "authentication", c.Authentication)
+	populate(objectMap, "awsCloudProfile", c.AwsCloudProfile)
 	populate(objectMap, "clusterExtensionIds", c.ClusterExtensionIDs)
 	populate(objectMap, "displayName", c.DisplayName)
+	populate(objectMap, "gcpCloudProfile", c.GcpCloudProfile)
 	populate(objectMap, "hostResourceId", c.HostResourceID)
 	populate(objectMap, "hostType", c.HostType)
 	populate(objectMap, "namespace", c.Namespace)
@@ -294,11 +327,17 @@ func (c *CustomLocationProperties) UnmarshalJSON(data []byte) error {
 		case "authentication":
 			err = unpopulate(val, "Authentication", &c.Authentication)
 			delete(rawMsg, key)
+		case "awsCloudProfile":
+			err = unpopulate(val, "AwsCloudProfile", &c.AwsCloudProfile)
+			delete(rawMsg, key)
 		case "clusterExtensionIds":
 			err = unpopulate(val, "ClusterExtensionIDs", &c.ClusterExtensionIDs)
 			delete(rawMsg, key)
 		case "displayName":
 			err = unpopulate(val, "DisplayName", &c.DisplayName)
+			delete(rawMsg, key)
+		case "gcpCloudProfile":
+			err = unpopulate(val, "GcpCloudProfile", &c.GcpCloudProfile)
 			delete(rawMsg, key)
 		case "hostResourceId":
 			err = unpopulate(val, "HostResourceID", &c.HostResourceID)
@@ -322,7 +361,7 @@ func (c *CustomLocationProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CustomLocationPropertiesAuthentication.
 func (c CustomLocationPropertiesAuthentication) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "type", c.Type)
 	populate(objectMap, "value", c.Value)
 	return json.Marshal(objectMap)
@@ -353,7 +392,7 @@ func (c *CustomLocationPropertiesAuthentication) UnmarshalJSON(data []byte) erro
 
 // MarshalJSON implements the json.Marshaller interface for type EnabledResourceType.
 func (e EnabledResourceType) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", e.ID)
 	populate(objectMap, "name", e.Name)
 	populate(objectMap, "properties", e.Properties)
@@ -396,7 +435,7 @@ func (e *EnabledResourceType) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type EnabledResourceTypeProperties.
 func (e EnabledResourceTypeProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "clusterExtensionId", e.ClusterExtensionID)
 	populate(objectMap, "extensionType", e.ExtensionType)
 	populate(objectMap, "typesMetadata", e.TypesMetadata)
@@ -431,7 +470,7 @@ func (e *EnabledResourceTypeProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type EnabledResourceTypePropertiesTypesMetadataItem.
 func (e EnabledResourceTypePropertiesTypesMetadataItem) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "apiVersion", e.APIVersion)
 	populate(objectMap, "resourceProviderNamespace", e.ResourceProviderNamespace)
 	populate(objectMap, "resourceType", e.ResourceType)
@@ -466,7 +505,7 @@ func (e *EnabledResourceTypePropertiesTypesMetadataItem) UnmarshalJSON(data []by
 
 // MarshalJSON implements the json.Marshaller interface for type EnabledResourceTypesListResult.
 func (e EnabledResourceTypesListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", e.NextLink)
 	populate(objectMap, "value", e.Value)
 	return json.Marshal(objectMap)
@@ -497,7 +536,7 @@ func (e *EnabledResourceTypesListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ErrorAdditionalInfo.
 func (e ErrorAdditionalInfo) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "info", &e.Info)
 	populate(objectMap, "type", e.Type)
 	return json.Marshal(objectMap)
@@ -528,7 +567,7 @@ func (e *ErrorAdditionalInfo) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ErrorDetail.
 func (e ErrorDetail) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "additionalInfo", e.AdditionalInfo)
 	populate(objectMap, "code", e.Code)
 	populate(objectMap, "details", e.Details)
@@ -571,7 +610,7 @@ func (e *ErrorDetail) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ErrorResponse.
 func (e ErrorResponse) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "error", e.Error)
 	return json.Marshal(objectMap)
 }
@@ -596,9 +635,40 @@ func (e *ErrorResponse) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type GCPCloudProfile.
+func (g GCPCloudProfile) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "organizationProject", g.OrganizationProject)
+	populate(objectMap, "singleProject", g.SingleProject)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type GCPCloudProfile.
+func (g *GCPCloudProfile) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", g, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "organizationProject":
+			err = unpopulate(val, "OrganizationProject", &g.OrganizationProject)
+			delete(rawMsg, key)
+		case "singleProject":
+			err = unpopulate(val, "SingleProject", &g.SingleProject)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", g, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type Identity.
 func (i Identity) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "principalId", i.PrincipalID)
 	populate(objectMap, "tenantId", i.TenantID)
 	populate(objectMap, "type", i.Type)
@@ -633,7 +703,7 @@ func (i *Identity) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type MatchExpressionsProperties.
 func (m MatchExpressionsProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "key", m.Key)
 	populate(objectMap, "operator", m.Operator)
 	populate(objectMap, "values", m.Values)
@@ -666,9 +736,44 @@ func (m *MatchExpressionsProperties) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type OrganizationProject.
+func (o OrganizationProject) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "excludedFolderIds", o.ExcludedFolderIDs)
+	populate(objectMap, "excludedProjectNumbers", o.ExcludedProjectNumbers)
+	populate(objectMap, "organizationId", o.OrganizationID)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type OrganizationProject.
+func (o *OrganizationProject) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", o, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "excludedFolderIds":
+			err = unpopulate(val, "ExcludedFolderIDs", &o.ExcludedFolderIDs)
+			delete(rawMsg, key)
+		case "excludedProjectNumbers":
+			err = unpopulate(val, "ExcludedProjectNumbers", &o.ExcludedProjectNumbers)
+			delete(rawMsg, key)
+		case "organizationId":
+			err = unpopulate(val, "OrganizationID", &o.OrganizationID)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", o, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type PatchableCustomLocations.
 func (p PatchableCustomLocations) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "identity", p.Identity)
 	populate(objectMap, "properties", p.Properties)
 	populate(objectMap, "tags", p.Tags)
@@ -703,7 +808,7 @@ func (p *PatchableCustomLocations) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type PatchableResourceSyncRule.
 func (p PatchableResourceSyncRule) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "properties", p.Properties)
 	populate(objectMap, "tags", p.Tags)
 	return json.Marshal(objectMap)
@@ -734,7 +839,7 @@ func (p *PatchableResourceSyncRule) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ProxyResource.
 func (p ProxyResource) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", p.ID)
 	populate(objectMap, "name", p.Name)
 	populate(objectMap, "type", p.Type)
@@ -769,7 +874,7 @@ func (p *ProxyResource) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Resource.
 func (r Resource) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", r.ID)
 	populate(objectMap, "name", r.Name)
 	populate(objectMap, "type", r.Type)
@@ -804,7 +909,7 @@ func (r *Resource) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ResourceSyncRule.
 func (r ResourceSyncRule) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", r.ID)
 	populate(objectMap, "location", r.Location)
 	populate(objectMap, "name", r.Name)
@@ -855,7 +960,7 @@ func (r *ResourceSyncRule) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ResourceSyncRuleListResult.
 func (r ResourceSyncRuleListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", r.NextLink)
 	populate(objectMap, "value", r.Value)
 	return json.Marshal(objectMap)
@@ -886,7 +991,7 @@ func (r *ResourceSyncRuleListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ResourceSyncRuleProperties.
 func (r ResourceSyncRuleProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "priority", r.Priority)
 	populate(objectMap, "provisioningState", r.ProvisioningState)
 	populate(objectMap, "selector", r.Selector)
@@ -925,7 +1030,7 @@ func (r *ResourceSyncRuleProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ResourceSyncRulePropertiesSelector.
 func (r ResourceSyncRulePropertiesSelector) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "matchExpressions", r.MatchExpressions)
 	populate(objectMap, "matchLabels", r.MatchLabels)
 	return json.Marshal(objectMap)
@@ -954,9 +1059,40 @@ func (r *ResourceSyncRulePropertiesSelector) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type SingleProject.
+func (s SingleProject) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "projectId", s.ProjectID)
+	populate(objectMap, "projectNumber", s.ProjectNumber)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type SingleProject.
+func (s *SingleProject) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "projectId":
+			err = unpopulate(val, "ProjectID", &s.ProjectID)
+			delete(rawMsg, key)
+		case "projectNumber":
+			err = unpopulate(val, "ProjectNumber", &s.ProjectNumber)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type SystemData.
 func (s SystemData) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populateTimeRFC3339(objectMap, "createdAt", s.CreatedAt)
 	populate(objectMap, "createdBy", s.CreatedBy)
 	populate(objectMap, "createdByType", s.CreatedByType)
@@ -1003,7 +1139,7 @@ func (s *SystemData) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type TrackedResource.
 func (t TrackedResource) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", t.ID)
 	populate(objectMap, "location", t.Location)
 	populate(objectMap, "name", t.Name)
@@ -1044,7 +1180,7 @@ func (t *TrackedResource) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-func populate(m map[string]interface{}, k string, v interface{}) {
+func populate(m map[string]any, k string, v any) {
 	if v == nil {
 		return
 	} else if azcore.IsNullValue(v) {
@@ -1054,7 +1190,7 @@ func populate(m map[string]interface{}, k string, v interface{}) {
 	}
 }
 
-func unpopulate(data json.RawMessage, fn string, v interface{}) error {
+func unpopulate(data json.RawMessage, fn string, v any) error {
 	if data == nil {
 		return nil
 	}